requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.32
asyncpg==0.29.0
redis==5.0.8
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
